backend/requirements.txt
alabaster==0.7.13
asgiref==3.8.1
Babel==2.15.0
backports.zoneinfo==0.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
cryptography==42.0.8
Django==4.2.13
django-allauth==0.63.3
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-extensions==3.2.3
djangorestframework==3.15.1
docutils==0.20.1
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
gspread==6.1.2
httplib2==0.22.0
idna==3.7
imagesize==1.4.1
importlib-metadata==7.1.0
jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.24.4
oauthlib==3.2.2
packaging==24.0
pandas==2.0.3
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
snowballstemmer==2.2.0
sphinx==7.1.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.5.0
typing-extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
zipp==3.19.0
